requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
numpy==2.1.3
scikit-learn==1.5.2
networkx==3.4.2
nltk==3.9.1
transformers==4.45.2
torch==2.4.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
safetensors==0.4.5
